backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
websockets>=12.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiogram>=3.0.0
